Forum/requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Modus==0.0.1
Flask-Script==2.0.6
Flask-Seeder==1.1.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-slugify==4.0.1
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
Werkzeug==0.16.1
WTForms==2.3.1
